refactor: move chat link utils to their own file
